Prevent the discovered write_after_free by disabling the committer before calling locallogout()
Rather than trying to cope with locallogout() clearing all the data structures when the call stack may still have an interest in those structures.
Assert added so we don't accidentally cause this again.
One unneeded commitNow() removed, the db->del() below will invoke it anyway, if needed.
A couple of warnings addressed (in 32 bit)

+++ tests/integration/SdkTest_test.cpp
@@ -3874,7 +3874,7 @@ TEST_F(SdkTest, SdkCloudraidStreamingSoakTest)
 
     int64_t filesize = getFilesize(filename2);
     std::ifstream compareDecryptedFile(filename2.c_str(), ios::binary);
-    ::mega::byte* compareDecryptedData = new ::mega::byte[filesize];
+    ::mega::byte* compareDecryptedData = new ::mega::byte[size_t(filesize)];
     compareDecryptedFile.read((char*)compareDecryptedData, filesize);
 
     m_time_t starttime = m_time();
